fix: more timeline efficiency improvements
